tmp105: change len and alorm to uint8_t
They were using only with very small integers, and they are sent/read as
bytes.  They can't become negative as far as I can see

+++ hw/tmp105.c
@@ -23,7 +23,7 @@
 
 typedef struct {
     i2c_slave i2c;
-    int len;
+    uint8_t len;
     uint8_t buf[2];
     qemu_irq pin;
 
@@ -32,7 +32,7 @@ typedef struct {
     int16_t temperature;
     int16_t limit[2];
     int faults;
-    int alarm;
+    uint8_t alarm;
 } TMP105State;
 
 static void tmp105_interrupt_update(TMP105State *s)
